add more functionality to LoadShedderManager

+++ flink-monitor/src/main/scala/berlin/bbdc/inet/mera/common/akka/AkkaMesseges.scala
@@ -4,4 +4,4 @@ final case class LoadShedderRegistration(id: String, address: String, port: Int)
 
 final case class ConfirmRegistration(id: String)
 
-final case class SendNewValue(value: Double)
+final case class SendNewValue(value: Int)
